refactor: remove layouts as not shareable
